// src/shared/api/generated/models/supportCaseAssignmentBatchResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { SupportCaseAssignmentBatchItemResponseDto } from "./supportCaseAssignmentBatchItemResponseDto";
import type { SupportCaseAssignmentBatchResponseDtoOutcome } from "./supportCaseAssignmentBatchResponseDtoOutcome";
import type { SupportCaseAssignmentBatchResponseDtoStatus } from "./supportCaseAssignmentBatchResponseDtoStatus";

export interface SupportCaseAssignmentBatchResponseDto {
  batchId: string;
  /**
   * @minimum 0
   * @maximum 50
   */
  failedCount: number;
  /**
   * @minimum 1
   * @maximum 50
   */
  itemCount: number;
  /** @maxItems 50 */
  items: SupportCaseAssignmentBatchItemResponseDto[];
  outcome: SupportCaseAssignmentBatchResponseDtoOutcome;
  /**
   * @minimum 0
   * @maximum 50
   */
  processedCount: number;
  status: SupportCaseAssignmentBatchResponseDtoStatus;
  /**
   * @minimum 0
   * @maximum 50
   */
  succeededCount: number;
}